src/main/java/com/paradigm_shifters/Randevu/model/Calculator.java
package com.paradigm_shifters.Randevu.model;

public class Calculator {

    public int square(int x) {
        return x * x;
    }

    public int countA(String word) {
        int count = 0;
        for (int i = 0; i < word.length(); i++) {
            if (word.charAt(i) == 'a' || word.charAt(i) == 'A') {
                count++;
            }
        }
        return count;
    }

    public int add(int a, int b) {
        return a + b;
    }

    public int subtract(int a, int b) {
        return a - b;
    }

    public int divide(int a, int b) {
        return a / b;
    }

    public int multiply(int a, int b) {
        return a * b;
    }
}
